+++ bin/nanocld/src/services/openapi.rs
@@ -296,6 +296,7 @@ impl Modify for VersionModifier {
     metric::list_metric,
     metric::create_metric,
     metric::inspect_metric,
+    metric::count_metric,
     // Process
     process::logs_processes,
     process::logs_process,